Add VHDL Langage
Not sure if you'll accept the VHDL in your ressources but it's a really interesting langage used to describe the behavior of hardware. It's used with FPGA, ASIC and because it's my job, I wanted to add it in the ressources. I often used this repo to learn some things and I'd like to contribute too.